// src/modules/emailNotification/emailNotification.enum.ts
export enum EmailNotificationType {
  StaffingRequestStatusPartiallyCovered = 'staffingRequest_statusPartiallyCovered',
  StaffingRequestDueDate24h = 'staffingRequest_dueDate24h',
  StaffingRequestDueDate48h = 'staffingRequest_dueDate48h',
  GeneralReport = 'generalReport',
  NPSReportGenerated = 'NPSReportGenerated',
  CasesReportGenerated = 'casesReportGenerated',
  GeneralType = 'general',
}